tv-chart: refactor - merge ChartContainer & PerformanceChart -> StrategyChart
